feat: Add assembly program and Makefile for OS boot menu
This commit adds an assembly program and a Makefile for an OS boot menu. The assembly program is located in `asm/os/src/main.asm` and includes code to set up segment registers, stack pointer, print a message, and halt the CPU. The Makefile is located in `asm/os/Makefile` and includes rules to build the assembly program into a binary file.